+++ sample/src/androidMain/kotlin/com/github/panpf/zoomimage/sample/ui/examples/compose/CoilZoomAsyncImageSample.kt
@@ -8,10 +8,10 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.tooling.preview.Preview
-import coil.request.ImageRequest.Builder
-import coil.size.Precision.INEXACT
+import coil3.request.ImageRequest.Builder
+import coil3.request.crossfade
+import coil3.size.Precision.INEXACT
 import com.github.panpf.sketch.fetch.newResourceUri
 import com.github.panpf.tools4a.toast.ktx.showShortToast
 import com.github.panpf.zoomimage.CoilZoomAsyncImage
@@ -27,11 +27,9 @@ fun CoilZoomAsyncImageSample(sketchImageUri: String) {
     ) { contentScale, alignment, state: ZoomState, scrollBar ->
         var myLoadState by remember { mutableStateOf<MyLoadState>(MyLoadState.None) }
         val context = LocalContext.current
-        val lifecycle = LocalLifecycleOwner.current.lifecycle
         val request = remember(key1 = sketchImageUri) {
             val model = sketchUri2CoilModel(context, sketchImageUri)
             Builder(context).apply {
-                lifecycle(lifecycle)
                 precision(INEXACT)
                 data(model)
                 crossfade(true)