+++ sample/src/androidMain/kotlin/com/github/panpf/zoomimage/sample/ui/photoalbum/compose/CoilListImage.kt
@@ -1,14 +1,23 @@
 package com.github.panpf.zoomimage.sample.ui.photoalbum.compose
 
+import android.content.Context
+import android.graphics.drawable.ColorDrawable
+import android.graphics.drawable.Drawable
+import androidx.annotation.ColorRes
+import androidx.annotation.DrawableRes
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
-import coil.compose.AsyncImage
-import coil.request.ImageRequest
+import androidx.core.content.res.ResourcesCompat
+import coil3.compose.AsyncImage
+import coil3.request.ImageRequest
+import coil3.request.crossfade
+import coil3.request.error
+import coil3.request.placeholder
+import com.github.panpf.sketch.drawable.IconDrawable
 import com.github.panpf.zoomimage.sample.R
-import com.github.panpf.zoomimage.sample.ui.photoalbum.view.iconDrawable
 import com.github.panpf.zoomimage.sample.util.sketchUri2CoilModel
 
 @Composable
@@ -21,11 +30,18 @@ fun CoilListImage(sketchImageUri: String, modifier: Modifier) {
         model = ImageRequest.Builder(context).apply {
             data(coilModel)
             placeholder(iconDrawable(context, R.drawable.ic_image_outline, R.color.placeholder_bg))
-            error(R.drawable.im_error)
+            error(iconDrawable(context, R.drawable.ic_error_baseline, R.color.placeholder_bg))
             crossfade(true)
         }.build(),
         modifier = modifier,
         contentScale = ContentScale.Crop,
         contentDescription = "photo",
+    )
+}
+
+fun iconDrawable(context: Context, @DrawableRes icon: Int, @ColorRes bg: Int): Drawable {
+    return IconDrawable(
+        icon = ResourcesCompat.getDrawable(context.resources, icon, null)!!,
+        background = ColorDrawable(ResourcesCompat.getColor(context.resources, bg, null)),
     )
 }