+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Translations: [简体中文](CHANGELOG_zh.md)
 * new: TileBitmap adds bitmapFrom attribute
 * change: zoomimage-view-sketch and zoomimage-compose-sketch modules upgraded to sketch4, while
   adding zoomimage-view-sketch3 and zoomimage-compose-sketch3 modules to continue to support sketch3
+* change: zoomimage-view-coil and zoomimage-compose-coil modules are upgraded to coil3, and new
+  The zoomimage-view-coil2 and zoomimage-compose-coil2 modules continue to support coil2
 
 ## v1.0.2